Move CLI for regenerating connection strength tables, move TableGroup
